requirements.txt
Flask==2.3.2
Pillow==9.5.0
qrcode==7.4.2
Werkzeug==2.3.4